xvmb 14.0 Helix

+++ supplemental/build-xbmc-ppa.sh
@@ -2,19 +2,19 @@
 # Build Script for custom XBMC RetroRig PPA
 #======================================================================== 
 #
-# Author:  Michael DeGuzis and Jens-Christian, 
-# Date:    20140831
-# Version: Patch Level 8
+# Author:  Michael DeGuzis and Jens-Christian Lache, 
+# Date:    20141004
+# Version: Patch Level 9
 # ========================================================================
 
 #define base version
-BASE=3:13.1
+BASE=3:14.0
 
 # define patch level
-PL=8
+PL=9
 
 #define xbmc branch to checkout
-BRANCH=gotham-retrorig-pl$PL
+BRANCH=retrorig-pl$PL
 
 
 clear
@@ -32,6 +32,14 @@ else
   echo ""
   echo "build target is source"
   echo ""
+fi
+
+#set build target
+if [[ -n "$1" ]]; then
+  arg0=$1
+else
+  # set up default
+  arg0=source
 fi
 
 sleep 2s
@@ -103,7 +111,14 @@ fi
 
 cd xbmc
 git checkout $BRANCH
-rm -rf .git 
+
+if [ "$arg0" == "source" ]; then
+  echo "removing .git folder"
+  rm -rf .git 
+else
+  echo "keeping .git folder"
+fi
+
 cp ~/RetroRig/Artwork/XBMC/Splash_retrorig.png ~/RetroRig/Artwork/XBMC/Splash.png media/
 cd ..
 tar cfj xbmc_$BASE.$PL.orig.tar.bz2 xbmc
@@ -141,7 +156,13 @@ echo "control"
 cp ~/RetroRig/supplemental/xbmc/control debian/
 
 echo "rules"
-cp ~/RetroRig/supplemental/xbmc/rules debian/
+if [ "$arg0" == "source" ]; then
+  cp ~/RetroRig/supplemental/xbmc/rules debian/
+else
+  cp ~/RetroRig/supplemental/xbmc/rules.localbuild debian/
+fi
+
+
 
 echo "setting up patches"
 rm -rf debian/patches/*
@@ -161,12 +182,8 @@ cd debian
 rm xbmc-addon-dev.install xbmc-eventclients-common.install xbmc-eventclients-dev.examples xbmc-eventclients-dev.install xbmc-eventclients-j2me.install xbmc-eventclients-j2me.manpages xbmc-eventclients-ps3.install xbmc-eventclients-ps3.manpages xbmc-eventclients-wiiremote.install  xbmc-eventclients-wiiremote.manpages xbmc-eventclients-xbmc-send.install xbmc-eventclients-xbmc-send.manpages xbmc-pvr-dev.install xbmc-screensaver-dev.install xbmc-visualization-dev.install
 cd ..
 
-if [[ -n "$1" ]]; then
-  arg0=$1
-else
-  # set up default
-  arg0=source
-fi
+#disable google test suite
+sed -i "s|configure_gtest=yes|configure_gtest=no|g" configure.in
 
 case "$arg0" in
   compile)
